requirements.txt
fastapi
uvicorn[standard]
tweepy
python-dotenv
requests
jinja2
python-multipart
websockets
